\name{RParcoords-package}
\alias{RParcoords-package}
\alias{RParcoords}
\docType{package}
\title{
Provides Parallel Coordinates for R Data frames.
}
\description{
Provides Parallel Coordinates for R Data frames.
}
\details{
\tabular{ll}{
Package: \tab RParcoords\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-16\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Bertjan Broeksema <broeksem@lippmann.lu>

Maintainer: Bertjan Broeksema <broeksem@lippmann.lu>
}
\references{}

\keyword{~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~}

\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
